refactor: merge chat logic into getAiContent
